按VS警告修改部分内容、实现VIStyleSheet样式表托管器、实现Windows上的系统颜色响应机制、为计时行为增加永续模式。YSP部分进行Chara抽象，恢复滤镜实现。

+++ Visindigo/Visindigo/VICore/VIMath.h
@@ -21,10 +21,10 @@ public:
 		}
 		return ret;
 	}
-	static float sin_0_1(float percent) {
+	static double sin_0_1(double percent) {
 		return (qSin(-PI/2 + PI*percent)+1.0)/2.0;
 	}
-	static float sin_0_1_0(float percent) {
+	static double sin_0_1_0(double percent) {
 		return qSin(percent * PI);
 	}
 };
